refactor: derived yage core to different mixins
