+++ demo/examples/growing-down-lines.html
@@ -92,7 +92,7 @@
             }
         });
 
-        new Yagr(chart4, {
+        const u = new Yagr(chart4, {
             title: {
                 text: 'On different scales',
             },
