+++ demo/examples/realtime.html
@@ -54,6 +54,9 @@
                     show: false,
                 }
             },
+            tooltip: {
+                sum: true,
+            },
             series: [
                 {data: timeline.map(id1), color: 'red', id: '1', interpolation: 'smooth'},
                 {data: timeline.map(id2), color: 'green', id: '2', interpolation: 'left'},
