+++ demo/examples/update-configs.html
@@ -1,0 +1,53 @@
+<html>
+    <head>
+        <title>Yagr</title>
+        <script src="../../dist/yagr.iife.js"></script>
+        <link rel="stylesheet" href="../../dist/index.css" />
+        <style>
+            .container {
+                margin-bottom: 26px;
+                height: 400px;
+                width: 100%;
+            }
+            .grid {
+                height: 400px;
+                display: flex;
+                width: 100%;
+                flex-direction: row;
+                justify-content: space-between;
+            }
+        </style>
+    </head>
+    <body>
+        <div class="row">
+            <div class="grid">
+                <div id="chart1" class="container"></div>
+            </div>
+            <button id="reset">Rest data</button>
+            <button id="add">Add data</button>
+        </div>
+   
+    <script>
+        const yagr = new Yagr(chart1, {
+            title: {text: 'Range bands'},
+            timeline: new Array(20).fill().map((_, i) => i * 1000),
+            series: [
+                {data: new Array(20).fill().map((_, i) => Math.random() * 6), color: 'red', id: '1'},
+                {data: new Array(20).fill().map((_, i) => Math.random() * 6), color: 'orange', id: '2'}
+            ],
+        });
+
+        window.reset.onclick = () => {
+            yagr.batch(() => {
+                yagr.config.timeline = new Array(20).fill().map((_, i) => i * 1000);
+                yagr.setSeries('1', {data: new Array(20).fill().map((_, i) => Math.random() * 6)});
+            });
+        };
+
+        window.add.onclick = () => {
+            const l = yagr.config.timeline[yagr.config.timeline.length - 1];
+            yagr.setSeries([l + 1000, l + 2000], [{id: '1', data: [Math.random() * 6, Math.random() * 20]}]);
+        };
+    </script>
+    </body>
+</html>
